+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [1.46.0] - 2021-06-17
+## [1.46.0] - 2021-06-17 [YANKED]
 
 ### Added
 - `regionalizationv2` query string.